requirements.txt
torch
pyyaml
pandas
matplotlib
datasets
pytest
